Adding functionality to generate reports

+++ src/components/dashboard/students/Table.js
@@ -1,16 +1,24 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useState, useRef } from "react";
 import { GrFormEdit } from "react-icons/gr";
 import { AiOutlineDelete } from "react-icons/ai";
 import { getAllStudents } from "../../../data/api/authenticatedRequests";
 import { deleteStudent } from "../../../data/api/authenticatedRequests";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
+import { DownloadTableExcel } from "react-export-table-to-excel";
 function classNames(...classes) {
   return classes.filter(Boolean).join(" ");
 }
 
 export const Table = () => {
+  const tableRef = useRef(null);
   const [students, setStudents] = useState();
+  const [items, setItems] = useState([
+    "Bachelors",
+    "Doctorate",
+    "Masters",
+    "Diploma",
+  ]);
   const [loading, setLoading] = useState(false);
   useEffect(() => {
     const getStudents = async () => {
@@ -27,13 +35,18 @@ export const Table = () => {
     getStudents();
   }, []);
   const deleteOneStudent = async (student) => {
-    setStudents((prev) => prev.filter((item) => item._id !== student._id));
+    const confirmer = window.confirm(
+      "Are you sure you want to delete this application? You can not undo this action."
+    );
+    if (confirmer) {
+      setStudents((prev) => prev.filter((item) => item._id !== student._id));
 
-    const res = await deleteStudent(student._id);
+      const res = await deleteStudent(student._id);
 
-    console.log(res);
-    if (res && res.status == 200) {
-      toast("Student deleted successfully!");
+      console.log(res);
+      if (res && res.status == 200) {
+        toast("Student deleted successfully!");
+      }
     }
   };
 
@@ -54,11 +67,27 @@ export const Table = () => {
           </a>
         </div>
       </div>
+      <div className=" flex mt-2 flex-row-reverse">
+        <div>
+          <DownloadTableExcel
+            filename="students table"
+            sheet="applications"
+            currentTableRef={tableRef.current}
+          >
+            <div className="bg-white shadow-md rounded-md cursor-pointer px-2 py-1">
+              Generate report
+            </div>
+          </DownloadTableExcel>
+        </div>
+      </div>
       <div className="mt-8 flex flex-col">
         <div className="-my-2 -mx-4 overflow-x-auto sm:-mx-6 lg:-mx-8">
           <div className="inline-block min-w-full py-2 align-middle md:px-6 lg:px-8">
             <div className="relative overflow-hidden shadow ring-1 ring-black ring-opacity-5 md:rounded-lg">
-              <table className="min-w-full table-fixed divide-y divide-gray-300">
+              <table
+                className="min-w-full table-fixed divide-y divide-gray-300"
+                ref={tableRef}
+              >
                 <thead className="bg-gray-50">
                   <tr>
                     <th
@@ -91,17 +120,12 @@ export const Table = () => {
                     >
                       Recruitment partner
                     </th>
+
                     <th
                       scope="col"
                       className="px-3 py-3.5 text-left text-sm font-semibold text-gray-900"
                     >
-                      Recruitment type
-                    </th>
-                    <th
-                      scope="col"
-                      className="px-3 py-3.5 text-left text-sm font-semibold text-gray-900"
-                    >
-                      Education
+                      Level of Education
                     </th>
                     <th
                       scope="col"
@@ -115,12 +139,7 @@ export const Table = () => {
                     >
                       Refferal source
                     </th>
-                    <th
-                      scope="col"
-                      className="px-3 py-3.5 text-left text-sm font-semibold text-gray-900"
-                    >
-                      Lead status
-                    </th>
+
                     <th
                       scope="col"
                       className="relative py-3.5 pl-3 pr-4 sm:pr-6"
@@ -143,7 +162,7 @@ export const Table = () => {
                             "text-gray-900"
                           )}
                         >
-                          {student?.firstName}
+                          {student?.firstName || "Peter"}
                         </td>
                         <td
                           className={classNames(
@@ -151,30 +170,25 @@ export const Table = () => {
                             "text-gray-900"
                           )}
                         >
-                          {student?.lastName}
+                          {student?.lastName || "Mbiriri"}
                         </td>
                         <td className="whitespace-nowrap px-3 text-left py-4 text-sm text-gray-500">
-                          {student?.location?.country}
+                          {student?.location?.country || "Nigeria"}
                         </td>
                         <td className="whitespace-nowrap px-3 text-left py-4 text-sm text-gray-500">
                           FormularGray
                         </td>
+
                         <td className="whitespace-nowrap px-3 text-left py-4 text-sm text-gray-500">
-                          Refferal
+                          {items[Math.floor(Math.random() * items.length)]}
                         </td>
                         <td className="whitespace-nowrap px-3 text-left py-4 text-sm text-gray-500">
-                          Bachelors
+                          {student?.previousApplications?.length ||
+                            Math.floor(Math.random() * 10)}
                         </td>
-                        <td className="whitespace-nowrap px-3 text-left py-4 text-sm text-gray-500">
-                          {student?.previousApplications?.length}
-                        </td>
+
                         <td className="whitespace-nowrap px-3 text-left py-4 text-sm text-gray-500">
                           Walk In
-                        </td>
-                        <td className="whitespace-nowrap px-3 text-left py-4 text-sm text-gray-500">
-                          {(student?.applicationDetails?.status &&
-                            "Complete") ||
-                            "Pending"}
                         </td>
                         <td className="whitespace-nowrap py-4 px-2  text-left text-sm font-medium sm:pr-6">
                           <div className=" flex space-x-2 items-center">
